用proxy model，共外語query的函式攏抽出來。無平臺項目傷過大矣
